Add date functionality, Refactor duke to kenobi, Move Parser into its own package and add ParseException
