--- src/main/java/duke/Duke.java
@@ -1,48 +1,0 @@
-package duke;
-
-import duke.command.Command;
-import duke.command.ExitCommand;
-import duke.util.Parser;
-import duke.util.Storage;
-import duke.util.TaskList;
-import duke.util.Ui;
-
-import java.util.Scanner;
-
-public class Duke {
-    TaskList tasks;
-    Storage storage;
-    Ui ui;
-
-    public Duke(String savePath) {
-        storage = new Storage(savePath);
-        tasks = storage.load();
-        ui = new Ui();
-    }
-
-    public void run() {
-        Scanner userInput = new Scanner(System.in);
-        Command nextCmd;
-        String response;
-
-        ui.greet();
-
-        while (true) {
-            nextCmd = Parser.parse(userInput.nextLine());
-            nextCmd.setData(tasks);
-            response = nextCmd.execute();
-            ui.say(response);
-
-            if (nextCmd instanceof ExitCommand) {
-                storage.save(tasks);
-                break;
-            }
-        }
-
-        userInput.close();
-    }
-
-    public static void main(String[] args) {
-        new Duke("../../../data/tasks.txt").run();
-    }
-}
