+++ src/main/java/kenobi/parser/ParseException.java
@@ -1,0 +1,7 @@
+package kenobi.parser;
+
+public class ParseException extends Exception {
+    public ParseException(String msg) {
+        super("Sorry, I don't understand this:" + msg);
+    }
+}
