+++ src/main/java/kenobi/task/ToDo.java
@@ -1,4 +1,4 @@
-package duke.task;
+package kenobi.task;
 
 public class ToDo extends Task {
   public ToDo(String name) {
