+++ src/main/java/kenobi/util/Ui.java
@@ -1,4 +1,4 @@
-package duke.util;
+package kenobi.util;
 
 import java.io.PrintWriter;
 import java.util.Scanner;
